@@ -2,10 +2,13 @@
   "name": "ugit",
   "version": "1.0.0",
   "description": "A Git client built with Electron and React",
-  "main": "src/main.js",
+  "main": "dist/main.js",
   "scripts": {
     "start": "electron-forge start",
-    "build": "webpack --mode development",
+    "build": "webpack --mode development && tsc src/main.ts --outDir dist --target ES2020 --module commonjs --moduleResolution node --esModuleInterop --allowSyntheticDefaultImports",
+    "build:prod": "webpack --mode production",
+    "build:main": "tsc src/main.ts --outDir dist --target ES2020 --module commonjs --moduleResolution node --esModuleInterop --allowSyntheticDefaultImports",
+    "typecheck": "tsc --noEmit",
     "package": "electron-forge package",
     "make": "electron-forge make"
   },
@@ -24,10 +27,12 @@
     },
     "files": [
       "main.js",
+      "main.ts",
       "index.html",
       "dist/**/*",
       "src/**/*",
-      "node_modules/**/*"
+      "node_modules/**/*",
+      "tsconfig.json"
     ],
     "win": {
       "target": "nsis",
@@ -53,11 +58,16 @@
     "@electron-forge/plugin-auto-unpack-natives": "^7.11.1",
     "@electron-forge/plugin-fuses": "^7.11.1",
     "@electron/fuses": "^1.8.0",
+    "@types/node": "^20.19.30",
+    "@types/react": "^19.2.10",
+    "@types/react-dom": "^19.2.3",
     "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
-    "electron": "^28.0.0",
+    "electron": "^40.1.0",
     "smart-buffer": "^4.2.0",
     "style-loader": "^3.3.3",
+    "ts-loader": "^9.5.4",
+    "typescript": "^5.9.3",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4"
   },
